requirements.txt
# Core
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Deep Learning
torch>=1.10.0
torchvision>=0.11.0

# Time Series
statsmodels>=0.13.0
prophet>=1.1.0
pmdarima>=1.8.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.4.0

# Utils
tqdm>=4.62.0
pyyaml>=6.0
joblib>=1.1.0

# Jupyter
jupyter>=1.0.0
ipython>=7.30.0

# Testing
pytest>=6.2.0
